Move FileIO to stryker4jvm
