+++ stryker4jvm/src/main/scala/stryker4jvm/reporting/reporters/JsonReporter.scala
@@ -1,13 +1,12 @@
 package stryker4jvm.reporting.reporters
 
 import cats.effect.IO
+import fs2.io.file.Path
 import mutationtesting.MutationTestResult
 import stryker4jvm.config.Config
-import stryker4jvm.core.files.FileIO
 import stryker4jvm.core.logging.Logger
+import stryker4jvm.files.FileIO
 import stryker4jvm.reporting.{FinishedRunEvent, IOReporter}
-
-import java.nio.file.Path
 
 class JsonReporter(fileIO: FileIO)(implicit log: Logger) extends IOReporter[Config] {
 
@@ -15,13 +14,13 @@ class JsonReporter(fileIO: FileIO)(implicit log: Logger) extends IOReporter[Conf
     import io.circe.syntax.*
     import mutationtesting.circe.*
     val json = report.asJson.noSpaces
-    IO(fileIO.createAndWrite(file, json))
+    fileIO.createAndWrite(file, json)
   }
 
   override def onRunFinished(runReport: FinishedRunEvent): IO[Unit] = {
     val resultLocation = fs2.io.file.Path.fromNioPath(runReport.reportsLocation) / "report.json"
 
-    writeReportJsonTo(resultLocation.toNioPath, runReport.report) *>
+    writeReportJsonTo(resultLocation, runReport.report) *>
       IO(log.info(s"Written JSON report to $resultLocation"))
   }
 }
