+++ stryker4jvm/src/main/scala/stryker4jvm/run/Stryker4jvmRunner.scala
@@ -5,9 +5,8 @@ import cats.effect.{IO, Resource}
 import fs2.io.file.Path
 import stryker4jvm.Stryker4jvm
 import stryker4jvm.config.{Config, ConfigReader, Console, Dashboard, Html, Json}
-import stryker4jvm.core.files.DiskFileIO
 import stryker4jvm.core.logging.Logger
-import stryker4jvm.files.{ConfigFilesResolver, FilesFileResolver, GlobFileResolver, MutatesFileResolver}
+import stryker4jvm.files.{ConfigFilesResolver, DiskFileIO, FilesFileResolver, GlobFileResolver, MutatesFileResolver}
 import stryker4jvm.logging.SttpLogWrapper
 import stryker4jvm.model.CompilerErrMsg
 import stryker4jvm.mutants.{Mutator, SupportedLanguageMutators}
